adjPval(maanova)	R Documentation

Generate FDR adjusted P values for F test result.

Description

This function takes a result object from matest and calcualte the FDR adjusted P values. The new P values will be appended to the input object as additional fields.

Usage


adjPval(matestobj, method=c("stepup","adaptive", "stepdown"))

Arguments

matestobj 	 An object of class matest, which is the result from matest.

method 	 The method for FDR control.

Value

An object of class matest with the following fields added for each F test:

adjPtab 	 FDR adjusted tabulated P values.

adjPvalperm 	 FDR adjusted permutation P values.

Author(s)

Hao Wu

Examples


data(paigen)
paigen <- createData(paigen.raw, n.rep=2)
model.noint.fix <- makeModel(data=paigen, formula=~Array+Dye+Spot+Strain+Diet)
# F-test strain effect
## Not run: 
test.strain.fix <- matest(paigen, model.noint.fix, term="Strain", n.perm=100,
        shuffle.method="resid", test.method=rep(1,4))
# make FDR adjusted P values
test.strain.fix <- adjPval(test.strain.fix)
## End(Not run)
# there will be new fields in test.strain.fix after this

[Package maanova version 1.2.0 Index]