arrayview(maanova)	R Documentation

View the layout of input data

Description

This function reconstructs the input data according to the micro array grid location structure and plots the data according to the user specified color map.

Dy default, it will plot the log ratios for 2-dye array and raw intensity for 1-dye array. It doesnot work for N-dye (N>2) array at this time.

Note that if user collapsed the replicates in creating the madata object (see createData), arrayview will be unavailable.

Usage


arrayview(object, ratio, array, colormap, onScreen=TRUE, ...)

Arguments

object 	 An object of class madata or rawdata.

ratio 	 The data to be ploted. The length of it must be equal to the length of the grid locations, .e.g, madata$row and madata$col. If ratio is a vector, there will be one plot. If ratio is a matrix, there will be one plot for each column. If ratio is not provided, link[maanova]{make.ratio} will be called to calculate the ratios from the original data.

array 	 A list of arrays to be plotted. This variable is only valid when ratio is not provided. Whenever ratio is provided, all columns in ratio will be plotted.

colormap 	 User specified color map. See colors for more detail.

... 	 Other parameters to be passed to image.

onScreen 	 A logical value to represent whether to display the plots on screen or not. If TRUE, x11() (in Unix/Windows) or macintosh (in Mac) will be called inside the function. Otherwise, it will plot the figure on the current device. Default is TRUE.

Author(s)

Hao Wu

Examples


## Not run: 
data(kidney)
############################
# arrayview plot on rawdata
############################
# arrayview raw data on screen
arrayview(kidney.raw, array=1)
graphics.off()
# arrayview raw data array 1 and 3 and output to postscript file
postscript(file="kidneyArrayview.ps")
arrayview(kidney.raw, array=c(1,3), onScreen=FALSE)
## End(Not run)

# once the replicates are collapsed,
# arrayview will be unavailable 
## Not run: data1 <- createData(kidney.raw, n.rep=2, avgreps=1)
## Not run: arrayview(data1) # get an error here

[Package maanova version 1.2.0 Index]