dyeswapfilter(maanova)	R Documentation

Gene filter for dye-swap experiment

Description

This function is used to flag the questionable spot in any kind of dye-swap experiment.

This function only works for 2-dye arrays.

Usage


dyeswapfilter(dataobj, r=4)

Arguments

dataobj 	 An object of class madata or rawdata.

r 	 A cut-off value for bad spot. The genes with log-ratio difference larger than r times standard deviation will be flagged.

Details

For each pair of dye-swap, the difference in log ratios (d) are computed. Then compute the IQR (interquartile range) of d and convert that to Standard Deviation by SD = IQR/1.35. Any gene with d larger than r times SD will be flagged.

Note that I assume in the input data object, the adjecent arrays is a dye-swap pair.

Value

An object of class rawdata or madata with the flag field created or updated.

Author(s)

Hao Wu

Examples


## Not run: 
data(kidney)
# riplot before filtering
riplot(kidney.raw, array=1)
# filter the gene
rawdata <- dyeswapfilter(kidney.raw)
# riplot again - some genes are highlighted
riplot(rawdata, array=1)
## End(Not run)

[Package maanova version 1.2.0 Index]