exprSet2Rawdata(maanova)	R Documentation

Convert an object of exprSet to an object of Rawdata

Description

This function converts an object of exprSet class, which is the main class for microarray data into an object of class "rawdata". It serves as a bridge between BioConductor and R/maanova.

Usage


exprSet2Rawdata(exprdata, ndye=1, trans.method="None") 

Arguments

exprdata 	 An object of class exprSet.

ndye 	 Number of dyes used in the experiment. Default is 1.

trans.method 	 A string for data transformation method used. Default is "None". This field is used only for display purpose. I suggest user to put something here as a reminder to yourself. For example if you run justGCRMA to get the exprdata, put "GCRMA" here.

Value

An object of class rawdata.

Author(s)

Hao Wu

[Package maanova version 1.2.0 Index]