geneprofile(maanova)	R Documentation

Expression plot for selected genes

Description

This function generate a plot with many lines. Each line represents a gene. The y-axis is the estimated expression level for the given factor from ANOVA model. The x-axis is for the levels of the give factor, e.g., different strains.

Usage


geneprofile(anovaobj, term, geneidx,
            col="blue", type="b", ylim, xlab, ylab, ...)

Arguments

anovaobj 	 An object of class maanova. It should be the result from fitmaanova.

term 	 The terms to be plotted.

geneidx 	 The index of genes to be plotted.

col 	 The color to be used in plot.

type 	 The line type.

ylim 	 Y-axis limit.

xlab 	 X-axis label.

ylab 	 Y-axis label.

... 	 Other parameters to be passed to plot.

Author(s)

Hao Wu

Examples


# load in data
data(paigen)
# make data object with rep 2
paigen <- createData(paigen.raw, 2)
# make an additive model
model.add.fix <- makeModel(data=paigen, formula=~Dye+Array+Strain+Diet)
# fit ANOVA model
anova.add <- fitmaanova(paigen, model.add.fix)
# test strain effect
## Not run: test.Strain.fix <- matest(paigen, model.add.fix, term="Strain", n.perm=100)
# volcano plot
idx <- volcano(test.Strain.fix)

# do gene profile for the selected genes
geneprofile(anova.add, "Strain", idx$idx.all)## End(Not run)

[Package maanova version 1.2.0 Index]