kidney.raw(maanova)	R Documentation

Kidney Data from CAMDA

Description

This is a 24-array double reference design. Six samples are compared to a reference with dye swapped and all arrays are duplicated. Flag for bad spots is included in the data.

Usage

data(kidney)

Format

An object of class rawdata.

Source

http://www.camda.duke.edu

References

Prichard CC, Hsu L, Delrow J and Nelson PS (2001), Project normal: defining normal variance in mouse gene expression, textit{PNAS}, 98:13266

Examples


data(kidney)

[Package maanova version 1.2.0 Index]