paigen.raw(maanova)	R Documentation

Data for a multiple factor 28-array expeirment

Description

This is a multiple factor 28-array experiment. The experiment is done in Beverly Paigen's Lab in The Jackson Lab. They took three strains of mice and feed them with two kind of diets. In that way you get six kind of mice. They picked two individuals in each group then you have totally 12 distinct mice. So in this experiment, you have strain, diet and biological replicates as the factors. You can test the effects from any factor or any combination of them.

Usage

data(paigen)

Format

An object of class rawdata.

Examples


data(paigen)

[Package maanova version 1.2.0 Index]