resiplot(maanova)	R Documentation

Residual plot for Micro Array Experiment

Description

This is the function to plot the residuals versus fitted value figure. Two channels, e.g., red and green, are drawn in seperate figures.

Usage


resiplot(madata, anovaobj, header)

Arguments

madata 	 An object of class madata.

anovaobj 	 An object of class maanova, which is the output from fitmaanova.

header 	 Optional. The title of the figure. The default figure title will be "Residual vs. Yhat plot".

Author(s)

Hao Wu

Examples


# load in Paigen's data
data(paigen)
# make data object with rep 2
paigen <- createData(paigen.raw, 2)
# Note that the data is normalized so normalization is skipped
# full model
model.full.fix <- makeModel(data=paigen,
      formula=~Dye+Array+Spot+Strain+Diet+Strain:Diet)
anova.full.fix <- fitmaanova(paigen, model.full.fix)
# residual plot
resiplot(paigen, anova.full.fix)

[Package maanova version 1.2.0 Index]