riplot(maanova)	R Documentation

Ratio intensity plot for 2-dye Micro Array experiment

Description

This function only works for 2-dye array at this time. It will plot the log-ratio (log2(R/G)) versus log-intensity (log2(R*G)/2) figure for Micro Array experiment. Ideal RI plot will be points scattered around the y=0 horizontal line.

This function works for both rawdata and madata. This function and arrayview assume the data field in rawdata is on the raw scale, and in madata is on log2 based scale. So if your rawdata is pre-transformed, you should not do riplot on the raw data.

Usage


riplot(object, title, array, color = "blue", highlight.flag = TRUE, 
    flag.color = "Red", idx.highlight, highlight.color = "Green", 
    rep.connect = FALSE, onScreen=TRUE) 

Arguments

object 	 An object of class madata or rawdata.

title 	 The title for figures. The default figure title is "RI plot for array number X". If the user wants to provide titles, be sure to provide a string array with the same number of elements as the number of arrays.

array 	 A list of arrays numbers for which you want to draw an RI plot.

color 	 The color for the points in scatter plot. Default is blue.

highlight.flag 	 A logical parameter to indicate whether to highlight the bad spots or not.

flag.color 	 The color for bad spots, default is red.

idx.highlight 	 A vector for highlighted spots other than bad spots.

highlight.color 	 The color for highlighted spots. Default is green.

rep.connect 	 A logical value to represent whether to connect the dots between the replicates or not.

onScreen 	 A logical value to represent whether to display the plots on screen or not. If TRUE, x11() (in Unix/Windows) or macintosh() (in Mac) will be called inside the function. Otherwise, it will plot the figure on the current device. Default is TRUE.

Note

This function will plot one figure for each array. So if you have many arrays, there will be many figures generated.

Author(s)

Hao Wu

Examples


## Not run: 
data(kidney)
############################
# riplot on rawdata
############################
# riplot raw data on screen
riplot(kidney.raw)
graphics.off()
# riplot raw data array 1 and 3 and output to postscript file
postscript(file="kidneyRIplot.ps")
riplot(kidney.raw, array=c(1,3), onScreen=FALSE)

############################
# RI plot on madata
############################
data1 <- createData(kidney.raw)
# do RI plot for all arrays
riplot(data1)
graphics.off()
## End(Not run)

[Package maanova version 1.2.0 Index]