subset.madata(maanova)	R Documentation

Subsetting Micro Array data objects

Description

Return subsets of an an object of class madata meeting given conditions.

Usage


subset.madata(x, arrays, genes, ...)

Arguments

x 	 An object of class madata. Read createData for details.

arrays 	 A vector specifying which arrays to keep or discard.

genes 	 A vector specifying which genes to keep or discard.

... 	 Ignored at this point.

Value

An object of class madata with specified arrays and genes.

Author(s)

Hao Wu

Examples


data(kidney)
# create data object with replicate
data <- createData(kidney.raw)
# take out array 1 and 2
smalldata <- subset(data, arrays=c(1,2))
# take out the all arrays except array 1
idx.array <- 1:data$n.array
smalldata <- subset(data,arrays=(idx.array[-1]))
# take out gene number 1 to 20
smalldata <- subset(data,genes=1:20)

[Package maanova version 1.2.0 Index]