varplot(maanova)	R Documentation

Variance component plot

Description

This function plots the density curve of each variance component of a result from fitmaanova.

If the input is from fixed model ANOVA, it will plot one curve for error variance component. If the input is from mixed model ANOVA, it will plot multiple curves, one for a random term (including error).

Usage


varplot(anovaobj)

Arguments

anovaobj 	 An object of class maanova.

Author(s)

Hao Wu

See Also

fitmaanova, density

Examples


# load in Paigen's data
data(paigen)
# make data object with rep 2
paigen <- createData(paigen.raw, 2)
## Not run: 
model.full.mix <- makeModel(data=paigen,
      formula=~Dye+Array+Spot+Strain+Diet+Strain:Diet+Sample,
      random=~Array+Spot+Sample)
anova.full.mix <- fitmaanova(paigen, model.full.mix, method="REML")
varplot(anova.full.mix)
## End(Not run)

[Package maanova version 1.2.0 Index]