write.madata(maanova)	R Documentation

Write Micro Array data to a TAB delimited simple text file

Description

This function is used to write the contents of an object of class madata or rawdata to a TAB delimited simple text file.

Usage


write.madata(madata, datafile="madata.txt", designfile="design.txt") 

Arguments

madata 	 The object to be output. It must be an object of class madata or rawdata.

datafile 	 The output file name for the data.

designfile 	 The output file name for the design file.

Author(s)

Hao Wu

Examples


# load in data
data(paigen)
# make data object
paigen <- createData(paigen.raw, 2)
# take out first 6 arrays
smallpaigen <- subset(paigen, array=1:6)
# write to file
## Not run: 
write.madata(smallpaigen, datafile="smallpaigen.txt",
      designfile="smallpaigendesign.txt")
## End(Not run)

[Package maanova version 1.2.0 Index]